backend/requirements.txt
flask
flask-jwt-extended
flask-cors
flask-mysqldb
flask-bcrypt
flasgger